+++ packages/react-ui/src/lib/components/index.ts
@@ -1,0 +1,2 @@
+export * from './ui';
+export * from './my-button/my-button';