+++ storybook/react/src/stories/components/page-sections/Background Image Hero.stories.tsx
@@ -11,7 +11,9 @@ const meta: Meta<typeof BackgroundImageHero> = {
     hero: {
       heading: 'Blog',
       subHeading: 'The latest from our team',
-      body: `We're always learning and growing. Here's what we've been up to lately.`,
+      body: [
+        `We're always learning and growing. Here's what we've been up to lately.`,
+      ],
     },
   },
 };