+++ storybook/react/src/stories/components/page-sections/Basic Hero.stories.tsx
@@ -10,7 +10,9 @@ const meta: Meta<typeof BasicHero> = {
     hero: {
       heading: 'Blog',
       subHeading: 'The latest from our team',
-      body: `We're always learning and growing. Here's what we've been up to lately.`,
+      body: [
+        `We're always learning and growing. Here's what we've been up to lately.`,
+      ],
     },
   },
 };