+++ storybook/react/src/stories/components/page-sections/Youtube With Content.stories.tsx
@@ -9,10 +9,13 @@ const meta: Meta<typeof YoutubeWithContent> = {
   args: {
     youtube: {
       id: '1xgP-dR4dzI',
-      placeholder: '/brand/youtube-placeholder.webp',
+      thumbnailUrl: '/brand/youtube-placeholder.webp',
     },
     content: {
       heading: 'A Suitably Epic Heading',
+      body: [
+        'ipsum lorem dolor sit amet consectetur adipiscing elit. Nullam nec purus et libero.',
+      ],
     },
   },
 };