+++ storybook/react/src/stories/ui/components/new-button.stories.ts
@@ -1,0 +1,14 @@
+import { NewButton } from '@/react-ui';
+import type { Meta, StoryObj } from '@storybook/react';
+
+const meta: Meta<typeof NewButton> = {
+  component: NewButton,
+  args: {
+    children: 'Click Me',
+  },
+  tags: ['autodocs'],
+};
+export default meta;
+type Story = StoryObj<typeof NewButton>;
+
+export const basic: Story = {};